requirements.txt
discordwebhook==1.0.2
pandas==1.3.4
numpy==1.22.0
scipy==1.7.3
tabulate
python-dotenv
scikit-learn
xgboost==1.5.2
lightgbm==3.3.1
joblib
numerapi==2.9.4